pyproject.toml
[project]
name = "knowledge-tracker"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic",
    "atproto",
    "tavily-python",
    "feedparser",
    "httpx",
    "beautifulsoup4",
    "pyyaml",
    "gitpython",
    "sentence-transformers",
    "torch",
    "scikit-learn",
    "apscheduler",
    "python-dotenv",
    "requests",
]

[project.scripts]
knowledge-tracker = "run:main"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "respx",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["."]
